requirements.txt
django
django-allauth
requests
jwt
dj-rest-auth
djangorestframework
djangorestframework-simplejwt
django-cors-headers
